--wip-- [skip ci]

+++ bin/build-all.sh
@@ -4,8 +4,14 @@ mkdir -p $ROOT_DIR/dist
 
 for project in $(ls_projs); do
     (
-        cd $project
-        hatch build
-        mv dist/* $ROOT_DIR/dist
+        ls $project/Dockerfile || echo not file
+        if [[ -f "$project/Dockerfile" ]] ; then
+            cd $ROOT_DIR
+            docker build --file  "$project/Dockerfile" --tag "monorepo/$(basename $project)" .
+        else
+            cd $project
+            hatch build
+            mv dist/* $ROOT_DIR/dist
+        fi
     )
 done
